Moving binprot and textprot packages under the protocol package
